src/requirements.txt
# Azure Functions runtime
azure-functions>=1.18.0,<2.0.0

# Azure SDKs
azure-servicebus>=7.11.4
azure-core>=1.30.0

# Config & Validation
pydantic>=2.0,<3.0.0
pydantic-settings>=2.0,<3.0.0
typing-extensions>=4.5.0

# Security & Code Quality
bandit>=1.7.5
safety>=2.3.1

# Testing
pytest>=7.0.0,<8.0.0

# Logging and telemetry
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
azure-monitor-opentelemetry>=1.0.0
